# Simplify native_to_unicode() & unicode_to_native()