@vthorsteinsson 2017-08-23
As the author of the original PR that is being modified here, I'm a bit sceptical about this change, as it degrades performance significantly under Python3. However, to avoid the lambda assignment to a variable, a simple change like the following would be easy to do:

```
if six.PY2:
  def native_to_unicode(s): return s if isinstance(s, unicode) else s.decode("utf8")
  def unicode_to_native(s): return s.encode("utf-8")
else:
  # No conversion required on Python >= 3
  def native_to_unicode(s): return s
  def unicode_to_native(s): return s
```